# schnorr: Use specialized types when signing.