--- a/dcrec/secp256k1/schnorr/error_test.go
+++ b/dcrec/secp256k1/schnorr/error_test.go
@@ -25,7 +25,6 @@ func TestErrorCodeStringer(t *testing.T) {
 		{ErrRegenerateRPoint, "ErrRegenerateRPoint"},
 		{ErrRegenSig, "ErrRegenSig"},
 		{ErrBadNonce, "ErrBadNonce"},
-		{ErrZeroSigS, "ErrZeroSigS"},
 		{ErrNonmatchingR, "ErrNonmatchingR"},
 		{ErrSigTooShort, "ErrSigTooShort"},
 		{ErrSigTooLong, "ErrSigTooLong"},