**This requires #2149**.

schnorr: Use specialized types when signing.

This modifies the function that produces signatures to use the `ModNScalar` and `FieldVal` types instead of big ints and also fleshes out the comments to include a reference to the signing algorithm specified in `README.md`.

This is work towards eventually using the specialized types throughout.

While the primary focus of this change is not optimization since signing happens infrequently enough that it is not in a performance hot path, as can be seen from the benchmark below, it does have the effect of making signing a bit faster.

The following benchmark shows a before and after comparison of producing a typical signature:

```
benchmark       old ns/op    new ns/op    delta
-------------------------------------------------
BenchmarkSign   54270        52622        -3.04%

benchmark       old allocs   new allocs   delta
-------------------------------------------------
BenchmarkSign   34           20           -41.18%

benchmark       old bytes    new bytes    delta
-------------------------------------------------
BenchmarkSign   1809         1088         -39.86%
```